fin-crawling-server/server/pyproject.toml
[tool.poetry]
name = "fin-crawling-fast-api"
version = "0.1.0"
description = ""
authors = ["tldrd0117 <tldrd0117@gmail.com>"]


[tool.poetry.scripts]
app = 'app.main'

[tool.poetry.dependencies]
python = "^3.7.3"
fastapi = "^0.64.0"
uvicorn = "^0.13.4"
pymitter = "^0.3.0"
watchdog = "^2.1.0"
pymongo = "^3.11.4"
selenium = "^3.141.0"
websockets = "^9.1"
python-multipart = "^0.0.5"
APScheduler = "^3.7.0"
aiofiles = "^0.7.0"
pandas = "^1.3.4"
openpyxl = "^3.0.9"
python-dotenv = "^0.19.2"
OpenDartReader = "^0.1.6"
bs4 = "^0.0.1"
aiohttp = "^3.8.1"
fake-useragent = "^0.1.11"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "^36.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.1"
flake8-mypy = "^17.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
